small fixed + distribution
git-svn-id: https://workrave.svn.sourceforge.net/svnroot/workrave/workrave/workrave/trunk@90 2b0e655b-6f86-497a-824e-95ef0e1e0d90

+++ gui/gtkmm/TimerPreferencesPanel.cc
@@ -290,7 +290,7 @@ TimerPreferencesPanel::on_preludes_maximum_toggled()
   int mp;
   if (has_max_prelude_cb->get_active())
     {
-      mp = max_prelude_adjustment.get_value();
+      mp = (int)max_prelude_adjustment.get_value();
     }
   else
     {
